+++ resources/js/tryitout.js
@@ -1,5 +1,21 @@
 window.abortControllers = {};
 
+function getCookie(name) {
+    if (!document.cookie) {
+        return null;
+    }
+
+    const cookies = document.cookie.split(';')
+        .map(c => c.trim())
+        .filter(c => c.startsWith(name + '='));
+
+    if (cookies.length === 0) {
+        return null;
+    }
+
+    return decodeURIComponent(cookies[0].split('=')[1]);
+}
+
 function tryItOut(endpointId) {
     document.querySelector(`#btn-tryout-${endpointId}`).hidden = true;
     document.querySelector(`#btn-executetryout-${endpointId}`).hidden = false;
@@ -85,7 +101,10 @@ function makeAPICall(method, path, body, query, headers, endpointId) {
         method,
         headers,
         body: method === 'GET' ? undefined : body,
-        signal: window.abortControllers[endpointId].signal
+        signal: window.abortControllers[endpointId].signal,
+        referrer: window.baseUrl,
+        mode: 'cors',
+        credentials: 'same-origin',
     })
         .then(response => Promise.all([response.status, response.text(), response.headers]));
 }
@@ -181,22 +200,11 @@ async function executeTryOut(endpointId, form) {
 
     const query = {};
     const queryParameters = form.querySelectorAll('input[data-component=query]');
-    queryParameters.forEach(el => _.set(query, el.name, el.value));
-
-    // Group radio buttons by their name, and then set the checked value from that group
-    Array.from(queryParameters)
-        .filter(el => el.type === "radio")
-        .reduce(
-            (entryMap, el) => entryMap.set(el.name, [...(entryMap.get(el.name) || []), el]),
-            new Map()
-        )
-        .forEach((v, k) => {
-            v.forEach(el => {
-                if (el.checked) {
-                    _.set(query, k, el.value);
-                }
-            });
-        });
+    queryParameters.forEach(el => {
+        if (el.type !== 'radio' || (el.type === 'radio' && el.checked)) {
+            _.set(query, el.name, el.value);
+        }
+    });
 
     let path = form.dataset.path;
     const urlParameters = form.querySelectorAll('input[data-component=url]');
@@ -220,7 +228,14 @@ async function executeTryOut(endpointId, form) {
         }
     }
 
-    makeAPICall(method, path, body, query, headers, endpointId)
+    let preflightPromise = Promise.resolve();
+    if (window.useCsrf && window.csrfUrl) {
+        preflightPromise = makeAPICall('GET', window.csrfUrl, {}, {}, {}, null).then(() => {
+            headers['X-XSRF-TOKEN'] = getCookie('XSRF-TOKEN');
+        });
+    }
+
+    return preflightPromise.then(() => makeAPICall(method, path, body, query, headers, endpointId))
         .then(([responseStatus, responseContent, responseHeaders]) => {
             handleResponse(endpointId, responseContent, responseStatus, responseHeaders)
         })