+++ src/Commands/Upgrade.php
@@ -11,25 +11,32 @@ class Upgrade extends Command
 
     protected $description = '';
 
+    public function newLine($count = 1)
+    {
+        // TODO Remove when Laravel 6 is no longer supported
+        $this->getOutput()->write(str_repeat("\n", $count));
+    }
+
     public function handle(): void
     {
         $oldConfig = config('scribe');
         $upgrader = Upgrader::ofConfigFile('config/scribe.php', __DIR__ . '/../../config/scribe.php')
-            ->dontTouch('routes')
+            ->dontTouch('routes', 'laravel.middleware', 'postman.overrides', 'openapi.overrides')
             ->move('interactive', 'try_it_out.enabled');
 
-        if ($this->option('dry-run')) {
-            $changes = $upgrader->dryRun();
-            if (empty($changes)) {
-                $this->info("No changes needed! Looks like you're all set.");
-                return;
-            }
+        $changes = $upgrader->dryRun();
+        if (empty($changes)) {
+            $this->info("No changes needed! Looks like you're all set.");
+            return;
+        }
 
-            $this->info('The following changes will be made to your config file:');
-            $this->newLine();
-            foreach ($changes as $change) {
-                $this->info($change["description"]);
-            }
+        $this->info('The following changes will be made to your config file:');
+        $this->newLine();
+        foreach ($changes as $change) {
+            $this->info($change["description"]);
+        }
+
+        if ($this->option('dry-run')) {
             return;
         }
 
@@ -42,10 +49,8 @@ class Upgrade extends Command
         }
 
         $this->newLine();
-        $this->info("✔ Upgraded your config to v3. Your old config is backed up at config/scribe.php.bak.");
-        $this->info("Please review to catch any mistakes.");
-        $this->warn("If you have any custom strategies or views, you should migrate those manually. See the migration guide at http://scribe.knuckles.wtf.");
-        $this->info("Don't forget to check out the release announcement for new features!");
+        $this->info("✔ Upgraded your config file. Your old config is backed up at config/scribe.php.bak.");
+        $this->info("Don't forget to check out the changelog or release announcement for new features!");
     }
 
 }