+++ src/Tools/Globals.php
@@ -4,7 +4,9 @@ namespace Knuckles\Scribe\Tools;
 
 class Globals
 {
-    public const SCRIBE_VERSION = '3.9.1';
+    public const SCRIBE_VERSION = '3.11.1';
 
     public static bool $shouldBeVerbose = false;
+
+    public static $beforeResponseCall;
 }